Correct the repair material for Elytra
This commit is in resolution of #1773 - in which Elytra were being repaired with leather. As of Minecraft 1.13 (in which the Phantom mob was added), Phantom Membranes are used to repair Elytra.

This commit simply creates a new repair item for Elytra, and registers it within AnvilRecipeMaker#getRepairRecipes

This commit can close #1773