Fix improper property use causing invalid dependency POM (#2854)
The archivesName property of BasePluginExtension is a Property<String>.
The default implementation of toString() for a Property is to print out
their name and owning object, and not the toString() of their contents.

This meant that passing archivesName to Node#appendNode as a value,
which calls toString() on it to make it into a String, results in e.g.
`extension 'base' property 'archivesName'` instead of the expected
value of e.g. `jei-1.19-common-api`.

The simple fix is to call Property#get() on archivesName to get its
actual value and pass that on as the node value.