Collate various potion reagents into single recipe (#4035)
JeiBrewingRecipe instances created by BrewingRecipeMakerCommon are given
a uid that combines purely the input potion type and the output potion
type.

This uid is used in equality checks, which means that additional recipes
for a potion with an existing recipe are completely ignored.

Instead of modifying the uid to include the reagent, this instead
collates existing reagents and, if the reagent of the "new" recipe
isn't already contained, replaces the existing recipe with a new one
that includes all previous reagents.